moved ArtificialTransport out of driver and into transport class, cleaned up boundary.py for missing smoothness parameters, updated init to take shape parameter
